Update CW gem, and add decorators for historical emissions
Given some differences between the South Africa platform and the main
platform I have added a few decorators to make it work on this project.

I would recommend revisiting this and making the gem a little more
generic so that it can cope with small differences. Where the
differences are not that small we should chat with the folks at wri to
make changes to the data sources.